+++ Userland/Libraries/LibJS/Runtime/Intl/ListFormat.h
@@ -13,7 +13,7 @@
 #include <AK/Vector.h>
 #include <LibJS/Runtime/Intl/AbstractOperations.h>
 #include <LibJS/Runtime/Object.h>
-#include <LibUnicode/Locale.h>
+#include <LibLocale/Locale.h>
 
 namespace JS::Intl {
