+++ Userland/Libraries/LibLocale/DateTimeFormat.h
@@ -12,8 +12,8 @@
 #include <AK/Time.h>
 #include <AK/Types.h>
 #include <AK/Vector.h>
+#include <LibLocale/Forward.h>
 #include <LibTimeZone/TimeZone.h>
-#include <LibUnicode/Forward.h>
 
 namespace Locale {
