+++ Userland/Libraries/LibLocale/NumberFormat.cpp
@@ -6,9 +6,9 @@
 
 #include <AK/CharacterTypes.h>
 #include <AK/Utf8View.h>
+#include <LibLocale/Locale.h>
+#include <LibLocale/NumberFormat.h>
 #include <LibUnicode/CharacterTypes.h>
-#include <LibUnicode/Locale.h>
-#include <LibUnicode/NumberFormat.h>
 
 #if ENABLE_UNICODE_DATA
 #    include <LibUnicode/UnicodeData.h>
