+++ Userland/Libraries/LibUnicode/CharacterTypes.cpp
@@ -10,8 +10,8 @@
 #include <AK/Types.h>
 #include <AK/Utf16View.h>
 #include <AK/Utf8View.h>
+#include <LibLocale/Locale.h>
 #include <LibUnicode/CharacterTypes.h>
-#include <LibUnicode/Locale.h>
 
 #if ENABLE_UNICODE_DATA
 #    include <LibUnicode/UnicodeData.h>
